requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
pillow==10.2.0
matplotlib==3.8.3
seaborn==0.13.2
#pickle5==0.0.12
#pickle5
pickle-mixin
# streamlit
# pandas
# numpy
# scikit-learn
# pillow
# matplotlib
# seaborn
# #pickle